# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "忆述光华 API",
    "description": "个人记忆整理与回顾后端。功能：认证（微信/手机号/JWT）、内容入库（照片/文字/语音）、四层事件聚合、描述性搜索（RAG）、同步（LWW）、AI 分类/画像/护栏。契约对齐《忆述光华_Sprint1规划.md》S1-02。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/search/image": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Search By Image Api",
        "description": "以图搜图（B2-4）：上传图片 → Qwen3-VL caption → image_vec 相似检索\n\n查询向量 = BGE-M3(caption)（caption 向量化方案）；返回与描述性搜索同构结果。",
        "operationId": "search_by_image_api_api_v1_search_image_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_search_by_image_api_api_v1_search_image_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SearchResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Search",
        "description": "描述性搜索主链路（API-003）：\nquery → LLM 改写/路由 → 多路召回（Qdrant dense+sparse RRF）→ payload filter\n→ bge-reranker 粗排 → qwen-flash 精排 → 溯源。\n\n验收：Top3≥70% + P95<3s（M1 门禁，RET-014/RET-018）。\n\n安全修复（审查 MAJOR）：个人记忆检索必须登录后可用（此前未鉴权）。",
        "operationId": "search_api_v1_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SearchResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/sync/push": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Sync Push",
        "operationId": "sync_push_api_v1_sync_push_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncPushRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SyncPushResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/sync/pull": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "Sync Pull",
        "operationId": "sync_pull_api_v1_sync_pull_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Since"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SyncPullResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/reconcile": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Sync Reconcile",
        "description": "端云对账（S5-04）：客户端本地快照 vs 云端权威 → 差异报告",
        "operationId": "sync_reconcile_api_v1_sync_reconcile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncReconcileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SyncReconcileResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/upload/init": {
      "post": {
        "tags": [
          "upload"
        ],
        "summary": "Init Upload",
        "operationId": "init_upload_api_v1_upload_init_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_init_upload_api_v1_upload_init_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/upload/chunk": {
      "put": {
        "tags": [
          "upload"
        ],
        "summary": "Upload Chunk",
        "operationId": "upload_chunk_api_v1_upload_chunk_put",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_chunk_api_v1_upload_chunk_put"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/upload/complete": {
      "post": {
        "tags": [
          "upload"
        ],
        "summary": "Complete Upload",
        "operationId": "complete_upload_api_v1_upload_complete_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_complete_upload_api_v1_upload_complete_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/upload/status": {
      "get": {
        "tags": [
          "upload"
        ],
        "summary": "Upload Status",
        "operationId": "upload_status_api_v1_upload_status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "upload_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Upload Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/upload/sts": {
      "get": {
        "tags": [
          "upload"
        ],
        "summary": "Upload Sts",
        "description": "客户端直传临时凭证（仅 cos 后端；STS 角色未就绪时给出降级提示）",
        "operationId": "upload_sts_api_v1_upload_sts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_SearchResult_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "OK"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SearchResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiResponse[SearchResult]"
      },
      "ApiResponse_SyncPullResult_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "OK"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SyncPullResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiResponse[SyncPullResult]"
      },
      "ApiResponse_SyncPushResult_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "OK"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SyncPushResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiResponse[SyncPushResult]"
      },
      "ApiResponse_SyncReconcileResult_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "OK"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SyncReconcileResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiResponse[SyncReconcileResult]"
      },
...
      "ApiResponse_dict_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "OK"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiResponse[dict]"
      },
...
      "Body_complete_upload_api_v1_upload_complete_post": {
        "properties": {
          "upload_id": {
            "type": "string",
            "title": "Upload Id"
          }
        },
        "type": "object",
        "required": [
          "upload_id"
        ],
        "title": "Body_complete_upload_api_v1_upload_complete_post"
      },
      "Body_init_upload_api_v1_upload_init_post": {
        "properties": {
          "client_upload_id": {
            "type": "string",
            "title": "Client Upload Id"
          },
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size"
          },
          "chunk_size": {
            "type": "integer",
            "title": "Chunk Size",
            "default": 8388608
          },
          "storage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage"
          }
        },
        "type": "object",
        "required": [
          "client_upload_id",
          "file_name",
          "file_size"
        ],
        "title": "Body_init_upload_api_v1_upload_init_post"
      },
      "Body_search_by_image_api_api_v1_search_image_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "查询图片（jpg/png，≤10MB）"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_search_by_image_api_api_v1_search_image_post"
      },
...
      "Body_upload_chunk_api_v1_upload_chunk_put": {
        "properties": {
          "upload_id": {
            "type": "string",
            "title": "Upload Id"
          },
          "chunk_index": {
            "type": "integer",
            "title": "Chunk Index"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "chunk_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Hash"
          }
        },
        "type": "object",
        "required": [
          "upload_id",
          "chunk_index",
          "file"
        ],
        "title": "Body_upload_chunk_api_v1_upload_chunk_put"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SearchQuery": {
        "properties": {
          "q": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Q",
            "description": "描述性查询（自然语言）"
          },
          "intent": {
            "type": "string",
            "enum": [
              "auto",
              "text",
              "image"
            ],
            "title": "Intent",
            "default": "auto"
          },
          "content_types": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Types"
          },
          "time_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time From"
          },
          "time_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time To"
          },
          "place": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place"
          },
          "tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          },
          "limit": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "q"
        ],
        "title": "SearchQuery"
      },
      "SearchResult": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "rewritten_query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rewritten Query"
          },
          "intent": {
            "type": "string",
            "title": "Intent"
          },
          "hits": {
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            },
            "type": "array",
            "title": "Hits"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms"
          },
          "degraded": {
            "type": "boolean",
            "title": "Degraded",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "query",
          "intent",
          "hits",
          "total",
          "latency_ms"
        ],
        "title": "SearchResult"
      },
...
      "SyncOp": {
        "properties": {
          "op_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Op Id"
          },
          "op_type": {
            "type": "string",
            "pattern": "^(upsert_field|delete)$",
            "title": "Op Type"
          },
          "entity_type": {
            "type": "string",
            "pattern": "^(content|event|profile)$",
            "title": "Entity Type",
            "default": "content"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id",
            "description": "实体 ID（uuid）"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field",
            "description": "upsert_field 时必填"
          },
          "value": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "integer"
              },
              {
                "type": "boolean"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At",
            "description": "客户端修改时间 ISO8601"
          }
        },
        "type": "object",
        "required": [
          "op_id",
          "op_type",
          "entity_id"
        ],
        "title": "SyncOp",
        "description": "客户端操作（op_id 全局唯一，幂等）"
      },
...
      "SyncPushRequest": {
        "properties": {
          "device_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Device Id"
          },
          "ops": {
            "items": {
              "$ref": "#/components/schemas/SyncOp"
            },
            "type": "array",
            "maxItems": 500,
            "title": "Ops"
          }
        },
        "type": "object",
        "required": [
          "device_id",
          "ops"
        ],
        "title": "SyncPushRequest"
      },
      "SyncPushResult": {
        "properties": {
          "applied": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Applied"
          },
          "conflicts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Conflicts"
          },
          "rejected": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Rejected",
            "default": []
          },
          "server_version": {
            "type": "integer",
            "title": "Server Version"
          }
        },
        "type": "object",
        "required": [
          "applied",
          "conflicts",
          "server_version"
        ],
        "title": "SyncPushResult"
      },
...
      "SyncReconcileRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SyncReconcileItem"
            },
            "type": "array",
            "maxItems": 5000,
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "SyncReconcileRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}